Split performance benchmark modules
