+++ scripts/perf/thresholds.ts
@@ -1,0 +1,11 @@
+export const DEFAULT_THRESHOLDS = {
+  avgMultiplier: 1.2,
+  p95Multiplier: 1.25,
+  avgAbsoluteFloorMs: 0.4,
+  p95AbsoluteFloorMs: 0.8,
+}
+
+export const CROSS_ENVIRONMENT_FLOORS = {
+  avgAbsoluteFloorMs: 1,
+  p95AbsoluteFloorMs: 2.5,
+}
